fix crash (oops!)

+++ common/src/main/java/com/st0x0ef/stellaris/common/data/screen/MoonPack.java
@@ -43,7 +43,7 @@ public class MoonPack extends SimpleJsonResourceReloadListener {
                         moon.texture(),
                         moon.name(),
                         (int) moon.distance(),
-                        (long) moon.period(),
+                        moon.period(),
                         (int) moon.width(),
                         (int) moon.height(),
                         PlanetSelectionScreen.findByNamePlanet(moon.parent()),
